+++ src/admin-auth/admin-auth.controller.ts
@@ -4,7 +4,6 @@ import {
   Body,
   HttpCode,
   HttpStatus,
-  Req,
 } from "@nestjs/common";
 import { AdminLoginDto } from "./dto/admin-login.dto";
 import { Token } from "./types/tokens.type";
