+++ src/admin/schema/user.schema.ts
@@ -1,6 +1,7 @@
 import { Prop, Schema, SchemaFactory } from "@nestjs/mongoose";
-import mongoose from "mongoose";
+import mongoose, { Document } from "mongoose";
 
+export type UserDocument = User & Document;
 
 @Schema({ timestamps: true, versionKey: false })
 export class User {
